+++ generators/typescript/sdk/CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.1] - 2024-05-14
+
+- Fix: If OAuth environment variables are specified, the `clientId` and `clientSecret` parameters
+  are optional.
+
+  ```ts
+  export declare namespace Client {
+    interface Options {
+        ...
+        clientId?: core.Supplier<string>;
+        clientSecret?: core.Supplier<string>;
+    }
+    ...
+  }
+  ```
+
 ## [0.18.0] - 2024-05-13
 
 - Feature: Add support for the OAuth client credentials flow. The new `OAuthTokenProvider` automatically